+++ src/hooks/use-queries.ts
@@ -12,11 +12,7 @@ import {
   ListTasksRequest,
   ListTrucksRequest,
   ListWorkShiftHoursRequest,
-  SalaryGroup,
 } from "generated/client";
-import { DateTime } from "luxon";
-
-const WORKING_TIME_PERIOD_START_DATE = DateTime.now().set({ day: 7, month: 1, year: 2024 });
 
 export const QUERY_KEYS = {
   SITES: "sites",
@@ -218,61 +214,6 @@ export const getFindHolidayQueryOptions = (holidayId: string, enabled = true) =>
     enabled: enabled,
     queryFn: async () => api.holidays.findHoliday({ holidayId }),
   });
-
-/**
- * Gets working period start and end datetime according to salary group and selected date
- *
- * @param salaryGroup Salary group of the employee
- * @param selectedDate Selected date
- * @returns Start and end date of the working period
- */
-export const getWorkingPeriodDates = (salaryGroup: SalaryGroup, selectedDate: Date) => {
-  const selectedDateTime = DateTime.fromJSDate(selectedDate);
-
-  return isOfficeOrTerminalGroup(salaryGroup)
-    ? getOfficeOrTerminalPeriod(selectedDateTime)
-    : getDriverPeriod(selectedDateTime);
-};
-
-const isOfficeOrTerminalGroup = (salaryGroup: string) => {
-  return salaryGroup === SalaryGroup.Office || salaryGroup === SalaryGroup.Terminal;
-};
-
-const getOfficeOrTerminalPeriod = (selectedDateTime: DateTime) => {
-  const midMonth = 16;
-
-  if (selectedDateTime.day < midMonth) {
-    const start = selectedDateTime.startOf("month").toJSDate();
-    const end = selectedDateTime.set({ day: 15 }).endOf("day").toJSDate();
-    return { start, end };
-  }
-
-  const start = selectedDateTime.set({ day: 16 }).startOf("day").toJSDate();
-  const end = selectedDateTime.endOf("month").endOf("day").toJSDate();
-  return { start, end };
-};
-
-const getDriverPeriod = (selectedDateTime: DateTime) => {
-  const fullWeeksFromStartDate = calculateFullWeeksFromStartDate(selectedDateTime);
-  const remainderRoundedUp = Math.ceil(fullWeeksFromStartDate % 2);
-  return calculatePeriod(fullWeeksFromStartDate - remainderRoundedUp);
-};
-
-const calculateFullWeeksFromStartDate = (selectedDateTime: DateTime) => {
-  return Math.floor(selectedDateTime.diff(WORKING_TIME_PERIOD_START_DATE, "weeks").weeks);
-};
-
-const calculatePeriod = (startWeekOffset: number) => {
-  const start = WORKING_TIME_PERIOD_START_DATE.plus({ weeks: startWeekOffset })
-    .set({ weekday: 7 })
-    .startOf("day")
-    .toJSDate();
-  const end = WORKING_TIME_PERIOD_START_DATE.plus({ weeks: startWeekOffset + 2 })
-    .set({ weekday: 6 })
-    .endOf("day")
-    .toJSDate();
-  return { start, end };
-};
 
 /**
  * Gets total results from headers
