+++ src/routes/working-hours_.$employeeId.work-shifts.tsx
@@ -15,27 +15,23 @@ import {
 } from "@mui/material";
 import { DatePicker } from "@mui/x-date-pickers/DatePicker";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { Outlet, createFileRoute } from "@tanstack/react-router";
+import { Outlet, createFileRoute, useRouter } from "@tanstack/react-router";
 import { api } from "api/index";
 import AggregationsTable from "components/working-hours/aggregations-table";
 import ChangeLog from "components/working-hours/change-log";
 import WorkShiftRow from "components/working-hours/work-shift-row";
 import WorkShiftsTableHeader from "components/working-hours/work-shifts-table-header";
 import { EmployeeWorkShift, SalaryGroup, WorkShiftHours, WorkType } from "generated/client";
-import {
-  QUERY_KEYS,
-  getListEmployeesQueryOptions,
-  getListTrucksQueryOptions,
-  getWorkingPeriodDates,
-} from "hooks/use-queries";
+import { QUERY_KEYS, getListEmployeesQueryOptions, getListTrucksQueryOptions } from "hooks/use-queries";
 import { t } from "i18next";
-import { DateTime, Interval } from "luxon";
+import { DateTime } from "luxon";
 import { FormProvider, useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import { toast } from "react-toastify";
 import { queryClient } from "src/main";
 import { Breadcrumb, EmployeeWorkHoursForm, EmployeeWorkHoursFormRow } from "src/types";
 import DataValidation from "src/utils/data-validation-utils";
+import TimeUtils from "src/utils/time-utils";
 import { z } from "zod";
 
 export const workShiftSearchSchema = z.object({
@@ -49,7 +45,7 @@ const addMissingWorkShiftRows = (
 ): EmployeeWorkHoursFormRow[] => {
   if (!workingPeriod) return formValues;
 
-  const allDatesInWorkingPeriod = eachDayOfInterval(workingPeriod.start, workingPeriod.end).map((date) =>
+  const allDatesInWorkingPeriod = TimeUtils.eachDayOfWorkingPeriod(workingPeriod.start, workingPeriod.end).map((date) =>
     date.toJSDate(),
   );
 
@@ -108,7 +104,7 @@ export const Route = createFileRoute("/working-hours_/$employeeId/work-shifts")(
     const employeeSalaryGroup =
       employees.find((employee) => employee.id === employeeId)?.salaryGroup ?? SalaryGroup.Driver;
 
-    const workingPeriodsForEmployee = getWorkingPeriodDates(employeeSalaryGroup, date?.toJSDate());
+    const workingPeriodsForEmployee = TimeUtils.getWorkingPeriodDates(employeeSalaryGroup, date?.toJSDate());
 
     const workShiftsData = await queryClient.ensureQueryData({
       queryKey: [QUERY_KEYS.WORK_SHIFTS, employeeId, "workShiftsData"],
@@ -146,7 +142,7 @@ export const Route = createFileRoute("/working-hours_/$employeeId/work-shifts")(
     const workShiftsDataWithWorkingPeriodDates = addMissingWorkShiftRows(
       employeeId,
       workShiftsData,
-      getWorkingPeriodDates(employeeSalaryGroup, date.toJSDate()),
+      TimeUtils.getWorkingPeriodDates(employeeSalaryGroup, date.toJSDate()),
     );
 
     const breadcrumbs: Breadcrumb[] = [
@@ -156,6 +152,8 @@ export const Route = createFileRoute("/working-hours_/$employeeId/work-shifts")(
     return { breadcrumbs, trucks, employees, workShiftsDataWithWorkingPeriodDates, employeeSalaryGroup };
   },
   component: WorkShifts,
+  errorComponent: () => <div>Error</div>,
+  pendingComponent: () => <div>Loading...</div>,
 });
 
 // Styled root component
@@ -230,19 +228,6 @@ const TableContainer = styled(Stack, {
   borderBottom: `1px solid ${theme.palette.divider}`,
 }));
 
-const eachDayOfInterval = (start: Date, end: Date): DateTime[] => {
-  const interval = Interval.fromDateTimes(
-    DateTime.fromJSDate(start).startOf("day"),
-    DateTime.fromJSDate(end).endOf("day"),
-  );
-
-  // Divide the interval into days and map them
-  return interval
-    .splitBy({ days: 1 })
-    .map((i) => i.start)
-    .filter((date): date is DateTime => date !== undefined);
-};
-
 function WorkShifts() {
   const { employees, workShiftsDataWithWorkingPeriodDates, employeeSalaryGroup } = Route.useLoaderData();
   const { t } = useTranslation();
@@ -251,159 +236,11 @@ function WorkShifts() {
   const queryClient = useQueryClient();
   const selectedDate = Route.useSearch({ select: (search) => search.date });
 
-  // const trucksQuery = useQuery(getListTrucksQueryOptions({}));
-  // const trucks = useMemo(() => trucksQuery.data?.trucks, [trucksQuery.data]);
-  // const employeesQuery = useQuery(getListEmployeesQueryOptions({}));
-  // const employees = useMemo(() => employeesQuery.data?.employees, [employeesQuery.data]);
-
-  // const employeeSalaryGroup = useMemo(
-  //   () => employees?.find((employee) => employee.id === employeeId)?.salaryGroup ?? SalaryGroup.Driver,
-  //   [employees, employeeId],
-  // );
-
-  // const workingPeriodsForEmployee = useMemo(() => {
-  //   const employeeSalaryGroup = employees?.find((employee) => employee.id === employeeId)?.salaryGroup;
-  //   if (!employeeSalaryGroup) return;
-  //   return getWorkingPeriodDates(employeeSalaryGroup, selectedDate.toJSDate());
-  // }, [employees, employeeId, selectedDate]);
-
-  // const workShiftsDataQuery = useQuery({
-  //   queryKey: [QUERY_KEYS.WORK_SHIFTS, employeeId, "workShiftsData"],
-  //   queryFn: async () => {
-  //     const [workShifts] = await api.employeeWorkShifts.listEmployeeWorkShiftsWithHeaders({
-  //       employeeId,
-  //       startedAfter: workingPeriodsForEmployee?.start,
-  //       startedBefore: workingPeriodsForEmployee?.end,
-  //     });
-
-  //     return await Promise.all(
-  //       workShifts.map<Promise<{ workShift: EmployeeWorkShift; workShiftHours: Record<WorkType, WorkShiftHours> }>>(
-  //         async (workShift) => {
-  //           const [workShiftHours] = await api.workShiftHours.listWorkShiftHoursWithHeaders({
-  //             employeeId,
-  //             employeeWorkShiftId: workShift.id,
-  //           });
-
-  //           return {
-  //             workShift: workShift,
-  //             workShiftHours: workShiftHours.reduce<Record<WorkType, WorkShiftHours>>(
-  //               (workShiftHoursRecord, singleWorkShiftHours) => {
-  //                 workShiftHoursRecord[singleWorkShiftHours.workType] = singleWorkShiftHours;
-  //                 return workShiftHoursRecord;
-  //               },
-  //               {} as Record<WorkType, WorkShiftHours>,
-  //             ),
-  //           };
-  //         },
-  //       ),
-  //     );
-  //   },
-  //   enabled: workingPeriodsForEmployee !== undefined,
-  // });
-
-  // const addMissingWorkShiftRows = useCallback(
-  //   (formValues: EmployeeWorkHoursFormRow[], workingPeriod: { start: Date; end: Date }): EmployeeWorkHoursFormRow[] => {
-  //     if (!workingPeriod) return formValues;
-
-  //     const allDatesInWorkingPeriod = eachDayOfInterval(workingPeriod.start, workingPeriod.end).map((date) =>
-  //       date.toJSDate(),
-  //     );
-
-  //     const existingWorkShiftDates = formValues.map((row) =>
-  //       DateTime.fromJSDate(row.workShift.date).startOf("day").toISODate(),
-  //     );
-
-  //     // Create empty rows for missing dates
-  //     const missingRows: EmployeeWorkHoursFormRow[] = allDatesInWorkingPeriod
-  //       .filter((date) => !existingWorkShiftDates.includes(DateTime.fromJSDate(date).toISODate()))
-  //       .map((missingDate) => ({
-  //         workShift: {
-  //           id: undefined, // No ID for missing work shifts
-  //           date: missingDate,
-  //           dayOffWorkAllowance: false,
-  //           absence: undefined,
-  //           perDiemAllowance: undefined,
-  //           approved: false,
-  //           notes: undefined,
-  //           startedAt: undefined,
-  //           endedAt: undefined,
-  //           truckIds: undefined,
-  //           employeeId,
-  //         } as EmployeeWorkShift,
-  //         // Create empty work shift hours for all work types
-  //         workShiftHours: Object.values(WorkType).reduce<Record<WorkType, WorkShiftHours>>(
-  //           (workShiftHours, workType) => {
-  //             workShiftHours[workType] = {
-  //               id: undefined, // No ID for missing work shift hours
-  //               employeeId,
-  //               workType,
-  //               calculatedHours: undefined,
-  //               actualHours: undefined,
-  //               employeeWorkShiftId: "", // No employee work shift ID for missing work shift hours
-  //             };
-  //             return workShiftHours;
-  //           },
-  //           {} as Record<WorkType, WorkShiftHours>,
-  //         ),
-  //       }));
-
-  //     const allRows = [...formValues, ...missingRows];
-
-  //     return allRows.sort(
-  //       (a, b) => DateTime.fromJSDate(a.workShift.date).toMillis() - DateTime.fromJSDate(b.workShift.date).toMillis(),
-  //     );
-  //   },
-  //   [employeeId],
-  // );
-
-  // const workShiftsDataWithWorkingPeriodDates = useMemo(
-  //   () =>
-  //     addMissingWorkShiftRows(
-  //       workShiftsDataQuery.data ?? [],
-  //       getWorkingPeriodDates(employeeSalaryGroup, selectedDate.toJSDate()),
-  //     ),
-  //   [addMissingWorkShiftRows, workShiftsDataQuery, employeeSalaryGroup, selectedDate],
-  // );
-
-  // useEffect(() => {
-  //   methods.reset(workShiftsDataWithWorkingPeriodDates);
-  // }, [workShiftsDataWithWorkingPeriodDates]);
-
   const methods = useForm<EmployeeWorkHoursForm>({
     defaultValues: workShiftsDataWithWorkingPeriodDates,
     values: workShiftsDataWithWorkingPeriodDates,
     mode: "onChange",
   });
-
-  // const getUpdatedWorkShiftsAndWorkShiftHours = (): [
-  //   updatedWorkShifts: EmployeeWorkShift[],
-  //   updatedWorkShiftHours: WorkShiftHours[],
-  // ] => {
-  //   const workShiftsToUpdate: EmployeeWorkShift[] = [];
-  //   const workShiftHoursToUpdate: WorkShiftHours[] = [];
-
-  //   const dirtyFormRows = Object.values(methods.getValues()).filter(
-  //     (_, index) => methods.getFieldState(`${index}`).isDirty,
-  //   );
-  //   console.log("dirtyFormRows", dirtyFormRows);
-  //   for (let i = 0; i < dirtyFormRows.length; i++) {
-  //     if (methods.getFieldState(`${i}.workShift`)?.isDirty) {
-  //       workShiftsToUpdate.push(methods.getValues(`${i}.workShift`));
-  //     }
-
-  //     const rowWorkTypes = Object.keys(methods.getValues(`${i}.workShiftHours`)) as WorkType[];
-
-  //     const dirtyWorkTypes = rowWorkTypes.filter(
-  //       (workType) => methods.getFieldState(`${i}.workShiftHours.${workType}`).isDirty,
-  //     );
-
-  //     workShiftHoursToUpdate.push(
-  //       ...dirtyWorkTypes.map((workType) => methods.getValues(`${i}.workShiftHours.${workType}`)),
-  //     );
-  //   }
-
-  //   return [workShiftsToUpdate, workShiftHoursToUpdate];
-  // };
 
   const getUpdatedWorkShiftsAndWorkShiftHours = (): [
     updatedWorkShifts: EmployeeWorkShift[],
@@ -423,8 +260,6 @@ function WorkShifts() {
 
     for (const i of dirtyFormRowIndices) {
       const row = methods.getValues(`${i}`);
-      console.log("ROW", row);
-      console.log("STANDBY", methods.getValues(`${i}.workShiftHours.STANDBY`));
 
       if (!row.workShift.id) {
         newRows.push(row);
@@ -451,39 +286,6 @@ function WorkShifts() {
 
     return [workShiftsToUpdate, workShiftHoursToUpdate, newRows];
   };
-
-  // const updateWorkShift = useMutation({
-  //   mutationFn: async () => {
-  //     const [updatedWorkShifts, updatedWorkShiftHours] = getUpdatedWorkShiftsAndWorkShiftHours();
-
-  //     await Promise.all(
-  //       updatedWorkShifts.map((workShift) =>
-  //         api.employeeWorkShifts.updateEmployeeWorkShift({
-  //           employeeId,
-  //           // biome-ignore lint/style/noNonNullAssertion: Work shift id is always defined
-  //           workShiftId: workShift.id!,
-  //           employeeWorkShift: workShift,
-  //         }),
-  //       ),
-  //     );
-
-  //     await Promise.all(
-  //       updatedWorkShiftHours.map((workShiftHours) =>
-  //         api.workShiftHours.updateWorkShiftHours({
-  //           // biome-ignore lint/style/noNonNullAssertion: Work shift id is always defined
-  //           workShiftHoursId: workShiftHours.id!,
-  //           workShiftHours: workShiftHours,
-  //         }),
-  //       ),
-  //     );
-  //   },
-  //   onError: () => toast.error(t("management.employees.errorToast")),
-  //   onSuccess: () => {
-  //     queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.WORK_SHIFTS] });
-  //     queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.WORK_SHIFT_HOURS] });
-  //     toast.success(t("workingHours.workingHourBalances.successToast"));
-  //   },
-  // });
 
   const updateWorkShift = useMutation({
     mutationFn: async () => {
@@ -539,14 +341,15 @@ function WorkShifts() {
         ),
       );
 
-      // Step 5: Invalidate queries to refresh form data
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.WORK_SHIFTS] });
       queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.WORK_SHIFT_HOURS] });
+
+      //window.location.reload();
 
       toast.success(t("workingHours.workingHourBalances.successToast"));
     },
     onError: () => {
-      toast.error(t("management.employees.errorToast"));
+      toast.error(t("workingHours.workingHourBalances.errorToast"));
     },
   });
 
@@ -629,7 +432,7 @@ function WorkShifts() {
   };
 
   const renderWorkingPeriodText = () => {
-    const workingPeriodDates = getWorkingPeriodDates(employeeSalaryGroup, selectedDate.toJSDate());
+    const workingPeriodDates = TimeUtils.getWorkingPeriodDates(employeeSalaryGroup, selectedDate.toJSDate());
     if (!workingPeriodDates) return null;
 
     const start = DateTime.fromJSDate(workingPeriodDates.start).setLocale("fi").toFormat("EEE dd.MM");
@@ -637,7 +440,7 @@ function WorkShifts() {
 
     return (
       <Typography variant="subtitle1">
-        {start} - {end}
+        {t("workingHours.workingHourBalances.payPeriod")}: {start} - {end}
       </Typography>
     );
   };
@@ -665,7 +468,6 @@ function WorkShifts() {
                 <TableContainer>
                   <WorkShiftsTableHeader />
                   {workShiftsDataWithWorkingPeriodDates.map((workShiftFormRow, index) => (
-                    // {workShiftsData.map(({ data: workShiftFormRow }, index) => (
                     <WorkShiftRow
                       key={`${index}_${workShiftFormRow.workShift.id}`}
                       onClick={() => navigate({ to: "work-shift-details", search: { date: selectedDate } })}
